site(diligence): give the deck the command that reproduces its own chart
The walkthrough never reached its headline. Step 5 runs
`spark benchmark run concurrency-sweep` at the gate's pinned shape — ISL
512 / OSL 320, C=1,4,8,16, one measured batch per cell — and the slide
immediately after it shows the published ladder: ISL 128 / OSL 1024,
C=1..128, three timed reps and a warmup, measured by
bench/ladder38/harness_w55_conc_ladder.py. A reader who ran every command
in order got numbers that cannot be compared to the chart in front of
them, and the deck named that driver in the fingerprint without ever
handing over a command to run it.

Add Step 5b: the campaign driver, at the exact shape published.json
records, all eight rungs. Every argument is derived — `--concs` from the
rung list the chart is drawn from, `--reps`/`--isl`/`--osl`/`--warmup`
from the workload block — so the command cannot drift from the figures
beside it, and a lost rung shortens the command as well as the ladder.
The driver itself defaults nothing (every argument is `required=True`),
which is what lets the command stand as the methodology.

Step 5 is retitled to say it measures the gate's workload, not the
chart's, so the two are no longer easy to confuse.

The slide also puts the two driver hashes side by side. The driver
sha256s its own source into every record it writes; the published Atlas
legs carry 1c77e1d8e9, and the copy in the tree today hashes 1f10d4887b
— it is neither of the two the manifest names. Behaviour matches (both
send presence_penalty/frequency_penalty at 0.0, the difference the
recorded pair was distinguished by), but the bytes do not, so the page
now tells a reader to record the hash they ran instead of quoting ours.
gen-ladder computes the tree's hash at build time rather than anyone
typing it, so it moves if the file is ever edited.

Act II is split at the measurement boundary to make room: Reproduce.svelte
keeps the reference frame and Steps 1-4, Ladder.svelte takes Step 5
onward. Slides register through context in DOM order, so the split needs
no manifest, and both files drop well under the 500-line rule that the
single file was sitting exactly on.
